Merge suites into single file (#11)
* Merge suites into single file

Before this change each test suite was declared in a separate file,
resulting in a bunch of files with a single meaningful line.

This commmit merges all suites declaration into |suite.js|.

It also drops "DASH PlayReady shaka Tests" and "DASH PlayReady dashjs
Tests" suites. PlayReady tests are now visible in "DASH shaka/dashjs Tests"
suite. There is a little point in keeping a separate suite for the sake
of 2 tests. Filtering of DRM tests on platforms which do not support it
should be achieved with the usage of profiles mechanism.