ARRISAPOL-3759 testSubtitles: keep static data
When subtitle test is started, one of the member arrays:

- this.content.subtitles.languages, or
- this.content.subtitles.testLanguages

is cloned and extended by 1 new entry.
The 'cloning' part wasn't the case for testLanguages
case; it was being used directly - and so when the test
using testLanguages was run in loop, testLanguages
array was being extended with each test run, making
the test take more time, until default timeout (30sec)
was being reached. The change corrects the problem.